adapt go
